pyproject.toml
[project]
name = "knowledgepitt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.67.0",
    "google-genai>=1.57.0",
    "groq>=1.0.0",
    "hf-xet>=1.2.0",
    "lightrag-hku[api]>=1.4.9.10",
    "onnxruntime>=1.23.2",
]
